Move useful test fixture conftest.py
We want to re-use this for other trap tests.
