+++ tests/trapobservers/logged_traps_test.py
@@ -1,11 +1,9 @@
-import ipaddress
 import logging
 from unittest.mock import Mock
 
 import pytest
 
-from zino.statemodels import DeviceState
-from zino.trapd import TrapMessage, TrapOriginator
+from zino.trapd import TrapMessage
 from zino.trapobservers.logged_traps import RestartTrapLogger
 
 
@@ -17,10 +15,3 @@ class TestRestartTrapLogger:
         with caplog.at_level(logging.INFO):
             observer.handle_trap(trap=trap)
             assert f"localhost: {trap_name}" in caplog.text
-
-
-@pytest.fixture
-def localhost_trap_originator():
-    addr = ipaddress.IPv4Address("127.0.0.1")
-    device = DeviceState(name="localhost", addresses=set((addr,)))
-    return TrapOriginator(address=addr, port=162, device=device)
